+++ policyengine_uk/variables/gov/gov_spending.py
@@ -44,4 +44,7 @@ class gov_spending(Variable):
         "nhs_budget_change",
         "education_budget_change",
         "other_public_spending_budget_change",
+        "tax_free_childcare",
+        "extended_childcare_entitlement",
+        "universal_childcare_entitlement",
     ]