+++ policyengine_uk/variables/household/income/income.py
@@ -168,6 +168,8 @@ class hbai_household_net_income(Variable):
         "basic_income",
         "cost_of_living_support_payment",
         "winter_fuel_allowance",
+        "tax_free_childcare",
+        # Reference for tax-free-childcare: https://assets.publishing.service.gov.uk/media/5e7b191886650c744175d08b/households-below-average-income-1994-1995-2018-2019.pdf
     ]
     subtracts = [
         "council_tax",
@@ -474,7 +476,7 @@ class capital_gains(Variable):
     definition_period = YEAR
     value_type = float
     unit = GBP
-    uprating = "calibration.programs.capital_gains.total"
+    uprating = "gov.obr.non_labour_income"
     adds = [
         "capital_gains_before_response",
         "capital_gains_behavioural_response",